requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
python-dotenv

